requirements.txt
asgiref==3.8.1
django==5.1.1
djangorestframework==3.15.2
markdown==3.7
sqlparse==0.5.3
